Add files via upload
# Business Problem

The Agricultural Pests Image Dataset is a collection of images of 12 different type of agricultural pests, namely Ants, Bees, Beetle, Catterpillar, Earthworms, Earwig, Grasshopper, Moth, Slug, Snail, Wasp, Weevil. This Dataset is for the Development and Evaluation of Machine Learning models for Pest Detection and Classification in Agriculture Settings. With 12 classes of pests, the Dataset provides a diverse range of images that cover a variety of shapes, colors, and sizes, making it suitable for training and testing algorithms to detect and classify pests in various scenarios.